add set loading to results table
